+++ modules/SofaTopologyMapping/SimpleTesselatedHexaTopologicalMapping.cpp
@@ -71,7 +71,7 @@ void SimpleTesselatedHexaTopologicalMapping::init()
                 toModel->addPoint(fromModel->getPX(i), fromModel->getPY(i), fromModel->getPZ(i));
             }
 
-            int pointIndex = pointMappedFromPoint.size();
+            size_t pointIndex = pointMappedFromPoint.size();
             Vector3 pA, pB, p;
 
             for (int i=0; i<fromModel->getNbHexahedra(); ++i)
@@ -236,7 +236,7 @@ void SimpleTesselatedHexaTopologicalMapping::init()
                 }
 
                 // points mapped from hexahedra
-                pointMappedFromHexa.push_back(pointIndex);
+                pointMappedFromHexa.push_back((int)pointIndex);
                 p = (p0+p1+p2+p3+p4+p5+p6+p7)/8;
                 toModel->addPoint(p[0], p[1], p[2]);
                 pointIndex++;
